refactor: 타입 관련 파일 위치 변경
